post-scan-actions/azure-python-slack-notification/slackNotification/requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions==1.7.2
azure-identity==1.6.0
